+++ e2e/ibrl_with_allocated_ip_test.go
@@ -92,7 +92,8 @@ func checkIBRLWithAllocatedIPPostConnect(t *testing.T, dn *TestDevnet, device *d
 		ones, _ := dzPrefixNet.Mask.Size()
 		allocatableBits := 32 - ones // number of host bits
 
-		expectedAllocatedClientIP, err := nextAllocatableIP(dzPrefixIP, allocatableBits, map[string]bool{})
+		// First IP is reserved for device tunnel endpoint (Loopback100 interface)
+		expectedAllocatedClientIP, err := nextAllocatableIP(dzPrefixIP, allocatableBits, map[string]bool{dzPrefixIP: true})
 		require.NoError(t, err)
 
 		dn.log.Info("--> Expected allocated client IP", "expectedAllocatedClientIP", expectedAllocatedClientIP, "deviceCYOAIP", device.CYOANetworkIP, "dzPrefix", device.DZPrefix)